revenue-risk-radar/requirements.txt
fastapi>=0.124.0
uvicorn[standard]>=0.38.0
streamlit>=1.41.0
python-dotenv==1.0.1
pydantic>=2.12.0
httpx>=0.28.0
certifi>=2026.2.25
